requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
monai==0.8.0
numpy==1.21.4
opencv-python==4.5.4.60
pandas==1.3.4
Pillow==8.4.0
pydicom==2.2.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==4.0.1
Werkzeug==2.0.2
